[PATCH] IDE: Fix request handling with ide-default & ATAPI
This fixes a bug that happens when a request gets to the IDE layer for a
drive using ide-default (that is with no subdriver attached), like a
Power Management request.  In this case, the core will wait for the
device status to match drive->read_stat, but that field contains by
default a value that is not suitable for ATAPI devices.  This patch
fixes it.

BKrev: 3f5e256cWu9LolxhKQdj7OEgveM2SQ