[PATCH] remap file pages MAP_NONBLOCK fix
From: Rajesh Venkatasubramanian <vrajesh@eecs.umich.edu>

The remap_file_pages system call with MAP_NONBLOCK flag does not
install file-ptes when the required pages are not found in the
page cache. Modify the populate functions to install file-ptes
if the mapping is non-linear and the required pages are not found
in the page cache.

Patch is for test4-mm6. Compiles and boots. Patch tested using the
programs at:

http://www-personal.engin.umich.edu/~vrajesh/linux/remap-file-pages/

BKrev: 3f5ed0c1llm6NnNwNXtPv-Z0IYzkwA