Merge bk://bk.arm.linux.org.uk/linux-2.6-pcmcia into home.osdl.org:/home/torvalds/v2.5/linux
2003/09/21 21:05:12+01:00 rmk
[PCMCIA] Fix deadlocks caused between PCMCIA card fix and device model

The problem was that the semaphore which prevents ds interfering with
the sleepy card initialisation (skt_sem in pccardd) is blocking insmod
of the socket driver.  However, the socket driver is being called with
the PCI bus semaphore held by the driver model.

pccardd in turn discovered a cardbus card (with skt_sem held), so it
is trying to add the PCI devices to the PCI bus, and this requires the
driver model to grab the PCI bus semaphore, but its already locked.

We move the class device register into pccardd so we get a natural
ordering between the ds socket initialisation and pccardd trying to
detect inserted cards.

We also fix a potential use-after-free caused by rmmod'ing the socket
driver before ds has shut down.

2003/09/17 16:57:12+01:00 rmk
[PCMCIA] Fix bug in PCMCIA resource management memory probing

This prevents us from probing past the memory areas which cardmgr
passes us, and also allows us to correctly remove areas which we
decide are not available for PCMCIA use.

BKrev: 3f6e189eezqBZBeAnPtdmlzwJwK_wg