Merge kroah.com:/home/greg/linux/BK/bleed-2.5 into kroah.com:/home/greg/linux/BK/i2c-2.6
2003/09/22 15:24:36-07:00 arvidjaar
[PATCH] I2C: sysfs sensor nameing inconsistency

> That's what you are going to have to set the name file to in the
> i2c_client structure, much like your patch did.  Then look at the
> different name files in each device directory to see what kind of device
> it is (chip, subclient, etc.)

OK attached patch sets all names to just chip name for chips themselves and
"chipname subclient" when subclient ios registered.

2003/09/22 15:11:24-07:00 greg
[PATCH] I2C: remove check_region usage and warning from i2c-sensor

2003/09/22 15:00:04-07:00 greg
[PATCH] I2C: move the remaining i2c bus drivers to drivers/i2c/busses

2003/09/22 14:21:09-07:00 greg
[PATCH] I2C: move the scx200* drivers to drivers/i2c/busses

2003/09/22 14:09:17-07:00 greg
[PATCH] I2C: move i2c-velleman driver to drivers/i2c/busses

2003/09/22 13:57:44-07:00 greg
I2C: move i2c-elektor.c driver to drivers/i2c/busses/

2003/09/22 13:44:13-07:00 greg
[PATCH] I2C: clean up the i2c-elv.c driver a bit

2003/09/22 13:36:52-07:00 greg
[PATCH] I2C: move i2c-elv.c driver to drivers/i2c/busses

2003/09/22 13:30:23-07:00 greg
[PATCH] I2C: clean up i2c-philips-par.c driver a bit

Fix bug when registering the i2c_adap as the structure was not set to 0.

2003/09/22 13:20:35-07:00 greg
[PATCH] I2C: move the i2c-philips-par driver to drivers/i2c/busses

2003/09/22 13:10:15-07:00 greg
[PATCH] I2C: fix up dependancies in the i2c/busses/Kconfig file

2003/09/22 13:06:04-07:00 greg
[PATCH] I2C: clean up i2c-prosavage.c driver

Remove direct memory accesses and link up device in the proper place in the
sysfs tree.

2003/09/22 12:46:46-07:00 greg
[PATCH] I2C: move i2c-prosavage.c driver to drivers/i2c/busses where it belongs.

2003/09/22 12:26:49-07:00 greg
I2C: clean up the drivers/i2c/Kconfig file

2003/09/22 11:42:40-07:00 greg
I2C: clean up the i2c chips Kconfig logic and help information

2003/09/22 11:15:16-07:00 greg
[PATCH] I2C: add the i2c-voodoo3 i2c bus driver

This is based on the lmsensor cvs version of the driver, but is cleaned
it up and ported it to 2.6.

2003/09/22 11:05:20-07:00 greg
[PATCH] I2C: add the i2c-savage4 i2c bus driver

This is based on the lmsensor cvs version of the driver, but is cleaned
it up and ported it to 2.6.

2003/09/22 11:03:30-07:00 greg
[PATCH] I2C: add the i2c-i810 i2c bus driver

This is based on the lmsensor cvs version of the driver, but is cleaned
it up and ported it to 2.6.

2003/09/15 16:54:37-07:00 rwhron
[PATCH] I2C: drivers/i2c version.h cleanup

remove unnecessary version.h from drivers/i2c.
test compiled with 2.6.0-test5-bk2 on x86.

2003/09/15 16:53:41-07:00 mporter
[PATCH] I2C: New PPC4xx I2C driver

This updates the current 2.6 PPC 4xx driver to the rewritten version
we have in the PPC development tree.  It is cleaner, has less bugs,
and has more features.

2003/09/15 16:44:07-07:00 greg
[PATCH] I2C: turn off debugging on the new sis i2c bus drivers.

Forgot to do this before...

2003/09/15 16:38:32-07:00 greg
[PATCH] I2C: clean up the i2c bus Kconfig menu and help texts.

2003/09/15 16:37:19-07:00 greg
[PATCH] I2C: add the i2c-via i2c bus driver

This is based on the lmsensor cvs version of the driver, but is cleaned
it up and ported it to 2.6.

2003/09/15 16:15:33-07:00 greg
[PATCH] I2C: add the i2c-sis630 i2c bus driver

This is based on the lmsensor cvs version of the driver, but is cleaned
it up and ported it to 2.6.

2003/09/15 13:46:00-07:00 greg
[PATCH] I2C: add the i2c-sis5595 i2c bus driver

This is based on the lmsensor cvs version of the driver, but is cleaned
it up and ported it to 2.6.

2003/09/12 17:04:05-07:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.5
into kroah.com:/home/greg/linux/BK/i2c-2.6

2003/09/12 16:58:01-07:00 greg
[PATCH] I2C: remove some usages of i2c_adapter.id as they are not used.

Seems i2c_adapter.id is only used in some video drivers, will work on
cleaning up that mess later...

2003/09/08 17:09:47-07:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.5
into kroah.com:/home/greg/linux/BK/i2c-2.6

2003/09/04 11:28:46-07:00 azarah
[PATCH] I2C: Fix conversion from milli volts in store_in_reg() for w83781d.c

I am not sure if it was a later patch from me that fixed in_* to display
milli volts in sysfs, or if it was a patch from Jan Dittmer, but the
conversion in the store_in_*() functions is wrong, and cause something
like:

2003/09/04 11:11:13-07:00 greg
Merge gregkh@kernel.bkbits.net:linux/i2c-2.5
into kroah.com:/home/greg/linux/BK/i2c-2.6

2003/08/27 15:32:59-07:00 hirofumi
[PATCH] DEVICE_NAME_SIZE/_HALF removal (I2C related, but fb stuff)

2003/08/27 15:32:44-07:00 hirofumi
[PATCH] DEVICE_NAME_SIZE/_HALF removal (I2C related, but v4l stuff)

2003/08/27 15:32:28-07:00 hirofumi
[PATCH] DEVICE_NAME_SIZE/_HALF removal (I2C stuff)

2003/08/27 15:29:33-07:00 mds
[PATCH] I2C: i2c-isa functionality

The "functionality" return was incorrectly removed for i2c-isa.c when porting to 2.5.
A driver that does nothing must return a 0 or else i2c-core assumes
0xffffffff (all functionality).

2003/08/27 15:24:53-07:00 greg
[PATCH] I2C: added new id for Radeon driver.

As requested by kronos@kronoz.cjb.net

BKrev: 3f6f796e-59P9pnR69cpNJBQwCzHHA