[PATCH] knfsd: NFS4XDR get rid of warning
There is a warning because of calling return with an argument on a void
function.  The root cause is putting a return statement in a macro (evil
side effect).

Change ENCODE_SEQID_OP_TAIL to be a more proper macro, change all
calling functions to return void.

BKrev: 3f6fae64Vj9L1CHb1m_nN8UFGiypXw