[PATCH] Minor K8 fix for oprofile
K8 can run with a 32bit kernel too.  If that's the case we have to tell
user space about it.  Remove the unneeded ifdef.

In that case the CPU is called "x86-64/hammer" even on 32bit.  This is
not strictly true on i386, but it's the named used by the current
oprofile user space for K8 performance counters and it would complicate
things to use different names between 32bit and 64bit.

BKrev: 3f70767f-m6LVLiFC1ueEc8dH53R9Q