[PATCH] fix AS hangs
Clears the next_arq cache if a request is repositioned and added as an alias
behind another.  Should fix the BUGs which Mary Edie Meredith has been
reporting.

Also, fixes a few random (but harmless) cases where last_merge was being set
and causing warnings.

When 2 requests are merged, and one is thrown away, move its alias list onto
the merged request! This is likely to be the cause of Mary's hangs.

BKrev: 3f74534cR__aA0JRbUd25v2vi9_MyA