Merge bk://ppc.bkbits.net/for-linus-ppc into home.osdl.org:/home/torvalds/v2.5/linux
2003/09/26 17:58:02-07:00 bernie
[PATCH] GCC 3.3.x/3.4 compatiblity fix in include/linux/init.h

GCC 3.4 miscompiles the kernel because it silently optimizes away
data placed in the .init.setup section by the __setup() macro.

__attribute__((unused)) does only avoid the warning, but doesn't
mark the data as being used.

Since GCC 3.3, __attribute__((used)) should be applied to such
variables.  The __attribute_used__ macro from linux/compiler.h already
takes care of compiler differences for us.

In this patch, I've gone a step further and proactively fixed that in
all places.

2003/09/26 17:55:31-07:00 sdake
[PATCH] fix kernel BUG using multipath

Thanks Matt and Jens for the debug help on the multipath problem.

This solves the problem and makes multipath work properly.

There are two types of "flags" that are used in a block io request,
bi_flags, and bi_rw.  bi_flags is used for flags to the block level
code, and bi_rw is used for flags to the low level device drivers.

The code in the multipath driver used the wrong flag in the wrong field.
In this case, the flag FASTFAIL (value 3) was being set to the bi_flags
field.  FASTFAIL is a hint to the low level driver that it should try to
fail out quickly.  Unfortunately, the value 3 is also BIO_SEG_VALID,
which is a flag to the block subsystem that the segments shouldn't be
recalculated.  The result was that the wrong field was set, telling the
block layer not to recalculate the segments resulting in phys and hw
segments of 0.  Not good.

BKrev: 3f750f54M2GjTfJWvqwE7QILC1i9rQ