[netdrvr sk98lin] use seq_file for /proc
Replace proc_read with seq_file interface which is much cleaner.

Also, instead of searching the list of devices looking for the
name which is slow and would break if interface name changes;
store a pointer to device in the proc entry, and retrieve
it from the information saved by single_open.

Formatting and other behaviours are retained.

BKrev: 3f755d92Sxc3Rc-zSmjpA9FJYlRXPg