[PATCH] select MII
The patch below switches MII to be select'ed instead of including it in
the Makefile.

Note that this patch requires a recent Linus' tree with the select CRC32
patch included.

diffstat output:

 drivers/net/Kconfig          |   11 +++++++++++
 drivers/net/Makefile         |   28 +++++++++++-----------------
 drivers/net/pcmcia/Kconfig   |    1 +
 drivers/net/tulip/Kconfig    |    1 +
 drivers/usb/net/Kconfig      |    2 ++
 drivers/usb/net/Makefile.mii |    6 ------
 6 files changed, 26 insertions(+), 23 deletions(-)


Tangential to the patch I observed a small problem (not fixed in the patch):
MII depends on NET_ETHERNET, but USB_PEGASUS and USB_USBNET depend
only on NET.

cu
Adrian

BKrev: 3f75710fvWLE1SEg7irfJ7kFPugOOg