[PATCH] for Alpha against 2.6.0-test6
Here is a very small set of fixes for Linux/axp, with a new one for
"sched_clock" just introduced.

Note that all but the arch/alpha/mm/init.c are required for a clean
build, since warnings->errors is still on.

The arch/alpha/mm/init.c stuff is from Jeff for MARVEL, and MUST be there
for any machines running the latest firmware...

BKrev: 3f7bbccbCjJTSpNRhF4v8ODDsn3UbQ