[PATCH] qla1280 locking update
This fixes the qla1280 driver to not try to release an unlocked spinlock
at boot time - as well as a couple of other minor mods.

I also changed the MMIO option to be dependant on X86_VISWS as it's
really a chipset problem and not something each and every driver should
be asking for.

Getting a global CONFIG_MMIO_IS_BROKEN option would probable be a win
here.

BKrev: 3f7bbe28jbqB62tVcvyZWIVRdZNdpQ