Merge kroah.com:/home/greg/linux/BK/bleed-2.5 into kroah.com:/home/greg/linux/BK/gregkh-2.6
2003/09/30 16:45:29-07:00 baldrick
[PATCH] USB speedtouch: neater check

This change was applied to 2.4, but not to 2.5.

2003/09/30 16:45:10-07:00 baldrick
[PATCH] USB speedtouch: reduce memory usage

Currently, incoming packets are reassembled in a sk_buff which is then
sent to the upper layers.  The sk_buff needs to be big enough to hold the
worst case packet since the size cannot be known in advance.  This would
not be so bad if the ATM layer paid attention to the mtu (usually 1500 bytes),
but for some reason it blithely ignores it and typically passes 9188 bytes
as the maximum size.  This means that at best 5/6 of the space in every
sk_buff is wasted.  So instead let's just allocate an assembly buffer (sarb)
which gets reused for every packet, and copy each assembled packet into
a minimally sized sk_buff before sending.

2003/09/30 16:44:52-07:00 baldrick
[PATCH] USB speedtouch: extra debug messages

2003/09/30 15:08:02-07:00 joe
[PATCH] USB: include/linux/usb.h

Reduce constant string space by reusing __FILE__

2003/09/30 15:02:31-07:00 wim
[PATCH] USB: problem with uhci-hcd in versions 2.6.0-test5 and 2.6.0-test6

I saw that you also reported problems with USB/uhci-hcd on your systems. Can you test
the following patch and see if it works now?

2003/09/30 15:02:06-07:00 mdharm-usb
[PATCH] USB: fix freecom.c

This patch for 2.6 fixes freecom.c -- an error in return-code
interpretation was introduced somewhere during 2.5.

This also adds a few more checks to try to catch commands that appear to
want to move data, but of length 0.  This is likely a bug in some higher
layer -- I'll bring it up on linux-scsi.

2003/09/30 15:01:46-07:00 stern
[PATCH] USB: unusual_devs.h update

2003/09/30 14:07:43-07:00 greg
Merge kroah.com:/home/linux/BK/bleed-2.5
into kroah.com:/home/linux/BK/gregkh-2.6

2003/09/25 16:33:12-07:00 amn3s1a
[PATCH] USB: New unusual_devs.h entry (Minolta DiMAGE E223 Digital Camera)

BKrev: 3f7dee0aayUUth-tVEC5wGWgt88WTw