[PATCH] unify drivers/Kconfig
It's a pain for architecture maintainers to keep the toplevel Kconfig
files in sync.  This addresses part of the problem by moving all the
standard driver includes to drivers/Kconfig.  As a bonus, it moves all
the device drivers together in a menu which tames menuconfig somewhat.

This patch only touches i386.  Other architectures can join the party
as soon as they feel ready to inflict all the choices on their userbase.

BKrev: 3f842e41l2SbQAU2DC2JUTLlj_HYDA