[PATCH] Fix compile on PPC
At the moment the aty128fb and matroxfb frame buffer drivers don't
compile on PPC.  In both cases the error is pretty trivial.  This
patch is the minimum change to fix the problems.

The _IOR() type problem fix in aty128fb.c depends on the fact that the
size of a pointer is the same as "size_t", so we don't have to use
_IOR_BAD() to keep the ioctl number the same.

BKrev: 3f856f62sfRj6KlGDTQ9hK3wGH9JZw