[PATCH] atp870u oops fix
This driver is taking uinitialised stack gunk from the pdev[] array and
feeding it into pci_read_config_byte() and crashing when modprobed with no
hardware present.

Fix it to not index past the initialised members of pdev[].

We don't know if this driver works.

BKrev: 3f96a00dHZRwX8ijoC-70AxHWjIwvQ