[PATCH] make printk more robust with "null" pointers
Expand printk's traditional handling of null pointers so that anything in the
first page is considered a null pointer.

This gives us better behaviour when someone (acpi..) accidentally prints a
string which is embedded in a struct, the pointer to which is null.

BKrev: 3f96a0b4LvlT5ql9z1mW3uPMTedgAg